move dashboard components into their own folder
