+++ frontend/app/components/dashboard/Widget.tsx
